+++ packages/core/src/vendors/PdfJs.d.ts
@@ -48,7 +48,9 @@ declare module 'pdfjs-dist' {
         data?: FileData;
         cMapUrl?: string;
         cMapPacked?: boolean;
+        httpHeaders?: Record<string, string | string[]>;
         url?: string;
+        withCredentials?: boolean;
     }
     function getDocument(params: GetDocumentParams): LoadingTask;